Extract array encoding step out of Json encoder and into  SpanEncoder
